feat: bad request if user has files related

+++ src/modules/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request, Req } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Request, Req, BadRequestException } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -12,29 +12,35 @@ export class UserController {
 
   @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body() createUserDto: CreateUserDto, @Request() req: ExpressRequest) {
-    return this.userService.create(createUserDto, req);
+  async create(@Body() createUserDto: CreateUserDto, @Request() req: ExpressRequest) {
+    return await this.userService.create(createUserDto, req);
   }
 
   @UseGuards(JwtAuthGuard)
   @Get()
-  findAll(@Request() req: ExpressRequest) {
-    return this.userService.findAll(req);
+  async findAll(@Request() req: ExpressRequest) {
+    return await this.userService.findAll(req);
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.userService.findOneById(id);
+  async findOne(@Param('id') id: string) {
+    return await this.userService.findOneById(id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.userService.update(id, updateUserDto);
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    return await this.userService.update(id, updateUserDto);
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string, @Request() req: ExpressRequest) {
+  async remove(@Param('id') id: string, @Request() req: ExpressRequest) {
+    const user = await this.userService.findOneById(id);
+
+    if (user.files.length > 0) {
+      throw new BadRequestException('No se puede eliminar un usuario porque tiene archivos asociados.');
+    }
+
     return this.userService.remove(id, req);
   }
 }
